fixed package names problem
