--- app-norx/src/build.gradle
@@ -1,72 +1,0 @@
-apply plugin: 'com.android.application'
-
-buildscript {
-    repositories {
-        mavenCentral()
-    }
-    dependencies {
-        // the latest version of the android-apt plugin
-        classpath 'com.neenbedankt.gradle.plugins:android-apt:1.8'
-    }
-}
-
-apply plugin: 'com.neenbedankt.android-apt'
-
-android {
-    compileSdkVersion 23
-    buildToolsVersion "23.0.3"
-
-    defaultConfig {
-        applicationId "com.alterego.stackoverflow.test"
-        minSdkVersion 19
-        targetSdkVersion 23
-        versionCode 1
-        versionName "1.1"
-        multiDexEnabled true
-    }
-    buildTypes {
-        release {
-            minifyEnabled  false
-            proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
-        }
-    }
-    packagingOptions {
-        exclude 'META-INF/services/javax.annotation.processing.Processor'
-    }
-}
-
-dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    compile 'com.android.support:appcompat-v7:23.4.0'
-    compile 'com.android.support:support-v4:23.4.0'
-    compile 'com.android.support:design:23.4.0'
-    compile 'com.android.support:cardview-v7:23.4.0'
-    compile 'com.android.support:recyclerview-v7:23.4.0'
-
-    compile 'joda-time:joda-time:2.4'
-
-    compile 'com.google.code.gson:gson:2.4'
-
-    compile 'io.reactivex:rxjava:1.1.6'
-    compile 'io.reactivex:rxandroid:1.2.1'
-
-    compile 'com.nostra13.universalimageloader:universal-image-loader:1.9.1'
-
-    compile 'com.squareup.okhttp:okhttp-urlconnection:1.6.0'
-    compile 'com.squareup.okhttp:okhttp:1.6.0'
-    compile 'com.squareup.retrofit:retrofit:1.6.1'
-
-    compile 'org.projectlombok:lombok:1.16.8'
-    compile 'de.greenrobot:greendao:1.3.7'
-
-    //ButterKnife Dependency
-    compile 'com.jakewharton:butterknife:8.1.0'
-    apt     'com.jakewharton:butterknife-compiler:8.1.0'
-    provided 'org.glassfish:javax.annotation:10.0-b28'
-
-    //Dagger Dependency for Injections
-    compile 'com.google.dagger:dagger:2.0.1'
-    apt "com.google.dagger:dagger-compiler:2.0.1"
-}
-
-
