+++ app-norx/src/main/java/com/alterego/stackoverflow/norx/test/MainActivity.java
@@ -1,0 +1,166 @@
+package com.alterego.stackoverflow.norx.test;
+
+
+import com.alterego.stackoverflow.norx.test.navigation.NavigationDrawerFragment;
+import com.alterego.stackoverflow.norx.test.search.SearchFragment;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
+import android.support.v4.widget.DrawerLayout;
+import android.support.v7.app.ActionBar;
+import android.support.v7.app.AppCompatActivity;
+import android.support.v7.widget.Toolbar;
+import android.text.Html;
+import android.view.Menu;
+import android.view.MenuItem;
+
+import solutions.alterego.stackoverflow.norx.test.R;
+
+public class MainActivity extends AppCompatActivity
+    implements NavigationDrawerFragment.NavigationDrawerCallbacks, OnFragmentInteractionListener, FragmentManager.OnBackStackChangedListener {
+
+    public static final String SAVED_TITLE = "saved_title";
+
+    private NavigationDrawerFragment drawerFragment;
+
+    private String title;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        MainApplication.component().inject(this);
+
+        setContentView(R.layout.activity_main);
+
+        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
+        setSupportActionBar(toolbar);
+
+        getSupportFragmentManager().addOnBackStackChangedListener(this);
+        shouldDisplayHomeUp();
+
+        drawerFragment = (NavigationDrawerFragment) getSupportFragmentManager().findFragmentById(R.id.navigation_drawer);
+
+        if (savedInstanceState != null) {
+            title = savedInstanceState.getString(SAVED_TITLE);
+        } else {
+            title = getTitle().toString();
+        }
+
+        drawerFragment.setUp(R.id.navigation_drawer, (DrawerLayout) findViewById(R.id.drawer_layout));
+    }
+
+    @Override
+    public void onNavigationDrawerItemSelected(int position) {
+        Fragment fragment_inst;
+
+        switch (position) {
+            case 0:
+                fragment_inst = new SearchFragment();
+                break;
+            default:
+                fragment_inst = new SearchFragment();
+        }
+
+        openFragmentInMainContainer(fragment_inst);
+    }
+
+    private void openFragmentInMainContainer(Fragment fragment) {
+        FragmentManager fragmentManager = getSupportFragmentManager();
+        fragmentManager.beginTransaction()
+            .replace(R.id.container, fragment)
+            .addToBackStack(null)
+            .commit();
+    }
+
+    public void onSectionAttached(int number) {
+        switch (number) {
+            case 1:
+                title = getString(R.string.title_section1);
+                break;
+            case 2:
+                title = getString(R.string.title_section3);
+                break;
+            case 3:
+                title = getString(R.string.title_section2);
+                break;
+        }
+    }
+
+
+    public void restoreActionBar() {
+        ActionBar actionBar = getSupportActionBar();
+//        actionBar.setDisplayHomeAsUpEnabled(true);
+//        actionBar.setNavigationMode(ActionBar.NAVIGATION_MODE_STANDARD);
+        actionBar.setDisplayShowTitleEnabled(true);
+        shouldDisplayHomeUp();
+        actionBar.setTitle(title);
+    }
+
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        if (!drawerFragment.isDrawerOpen()) {
+            // Only show items in the action bar relevant to this screen
+            // if the drawer is not showing. Otherwise, let the drawer
+            // decide what to show in the action bar.
+            getMenuInflater().inflate(R.menu.main, menu);
+            restoreActionBar();
+            return true;
+        }
+        return super.onCreateOptionsMenu(menu);
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+        // Handle action bar item clicks here. The action bar will
+        // automatically handle clicks on the Home/Up button, so long
+        // as you specify a parent activity in AndroidManifest.xml.
+        int id = item.getItemId();
+//        if (id == R.id.action_settings) {
+//            return true;
+//        }
+        return super.onOptionsItemSelected(item);
+    }
+
+    @Override
+    public void onRequestOpenFragment(Fragment fragment, String title) {
+        setActionBarTitle(title);
+        openFragmentInMainContainer(fragment);
+    }
+
+    @Override
+    public void setActionBarTitle(String title) {
+        if (title != null) {
+            this.title = title;
+        }
+
+        if (getSupportActionBar() != null) {
+            getSupportActionBar().setTitle(Html.fromHtml(this.title));
+        }
+    }
+
+    @Override
+    protected void onSaveInstanceState(Bundle outState) {
+        super.onSaveInstanceState(outState);
+        outState.putString(SAVED_TITLE, title);
+    }
+
+    @Override
+    public void onBackStackChanged() {
+        shouldDisplayHomeUp();
+    }
+
+    public void shouldDisplayHomeUp() {
+        //Enable Up button only  if there are entries in the back stack
+        boolean canback = getSupportFragmentManager().getBackStackEntryCount() > 0;
+        getSupportActionBar().setDisplayHomeAsUpEnabled(canback);
+    }
+
+    @Override
+    public boolean onSupportNavigateUp() {
+        //This method is called when the up button is pressed. Just the pop back stack.
+        getSupportFragmentManager().popBackStack();
+        return true;
+    }
+}
