+++ app-norx/src/main/java/com/alterego/stackoverflow/norx/test/OnFragmentInteractionListener.java
@@ -1,0 +1,8 @@
+package com.alterego.stackoverflow.norx.test;
+
+import android.support.v4.app.Fragment;
+
+public interface OnFragmentInteractionListener {
+    public void setActionBarTitle(String title);
+    public void onRequestOpenFragment(Fragment fragment, String title);
+}
