--- app-norx/src/main/java/com/alterego/stackoverflow/test/MainActivity.java
@@ -1,180 +1,0 @@
-package com.alterego.stackoverflow.test;
-
-import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.widget.DrawerLayout;
-import android.support.v7.app.ActionBar;
-import android.support.v7.app.ActionBarActivity;
-import android.support.v7.widget.Toolbar;
-import android.text.Html;
-import android.view.Menu;
-import android.view.MenuItem;
-
-import com.alterego.flickr.app.test.R;
-import com.alterego.stackoverflow.test.navigation.NavigationDrawerFragment;
-import com.alterego.stackoverflow.test.search.SearchFragment;
-
-
-public class MainActivity extends ActionBarActivity
-        implements NavigationDrawerFragment.NavigationDrawerCallbacks, OnFragmentInteractionListener, FragmentManager.OnBackStackChangedListener {
-
-    public static final String SAVED_TITLE = "saved_title";
-
-    /**
-     * Fragment managing the behaviors, interactions and presentation of the navigation drawer.
-     */
-    private NavigationDrawerFragment mNavigationDrawerFragment;
-
-    /**
-     * Used to store the last screen title. For use in {@link #restoreActionBar()}.
-     */
-    private String mTitle;
-    private SettingsManager mSettingsManager;
-//    private ActionBar mActionBar;
-
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        System.setProperty("org.joda.time.DateTimeZone.Provider", "com.alterego.stackoverflow.test.helpers.FastDateTimeZoneProvider");
-
-        setContentView(R.layout.activity_main);
-
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
-        setSupportActionBar(toolbar);
-//        getSupportActionBar().setDisplayHomeAsUpEnabled(true);
-//        getSupportActionBar().setHomeButtonEnabled(true);
-
-        mSettingsManager = MainApplication.getMainApplication().getSettingsManager();
-        mSettingsManager.setParentActivity(this);
-        mSettingsManager.getLogger().info("MainActivity onCreate");
-        getSupportFragmentManager().addOnBackStackChangedListener(this);
-        shouldDisplayHomeUp();
-
-        mNavigationDrawerFragment = (NavigationDrawerFragment)
-                getSupportFragmentManager().findFragmentById(R.id.navigation_drawer);
-
-        if (savedInstanceState != null) {
-            mTitle = savedInstanceState.getString(SAVED_TITLE);
-        } else
-            mTitle = getTitle().toString();
-
-        // Set up the drawer.
-        mNavigationDrawerFragment.setUp(
-                R.id.navigation_drawer,
-                (DrawerLayout) findViewById(R.id.drawer_layout));
-    }
-
-    @Override
-    public void onNavigationDrawerItemSelected(int position) {
-        Fragment fragment_inst;
-
-        switch (position) {
-            case 0:
-                fragment_inst = new SearchFragment();
-                break;
-            default:
-                fragment_inst = new SearchFragment();
-        }
-
-        openFragmentInMainContainer(fragment_inst);
-    }
-
-    private void openFragmentInMainContainer(Fragment fragment) {
-        FragmentManager fragmentManager = getSupportFragmentManager();
-        fragmentManager.beginTransaction()
-                .replace(R.id.container, fragment)
-                .addToBackStack(null)
-                .commit();
-    }
-
-    public void onSectionAttached(int number) {
-        switch (number) {
-            case 1:
-                mTitle = getString(R.string.title_section1);
-                break;
-            case 2:
-                mTitle = getString(R.string.title_section3);
-                break;
-            case 3:
-                mTitle = getString(R.string.title_section2);
-                break;
-        }
-    }
-
-
-    public void restoreActionBar() {
-        ActionBar actionBar = getSupportActionBar();
-//        actionBar.setDisplayHomeAsUpEnabled(true);
-//        actionBar.setNavigationMode(ActionBar.NAVIGATION_MODE_STANDARD);
-        actionBar.setDisplayShowTitleEnabled(true);
-        shouldDisplayHomeUp();
-        actionBar.setTitle(mTitle);
-    }
-
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        if (!mNavigationDrawerFragment.isDrawerOpen()) {
-            // Only show items in the action bar relevant to this screen
-            // if the drawer is not showing. Otherwise, let the drawer
-            // decide what to show in the action bar.
-            getMenuInflater().inflate(R.menu.main, menu);
-            restoreActionBar();
-            return true;
-        }
-        return super.onCreateOptionsMenu(menu);
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-        // Handle action bar item clicks here. The action bar will
-        // automatically handle clicks on the Home/Up button, so long
-        // as you specify a parent activity in AndroidManifest.xml.
-        int id = item.getItemId();
-//        if (id == R.id.action_settings) {
-//            return true;
-//        }
-        return super.onOptionsItemSelected(item);
-    }
-
-    @Override
-    public void onRequestOpenFragment(Fragment fragment, String title) {
-        setActionBarTitle(title);
-        openFragmentInMainContainer(fragment);
-    }
-
-    @Override
-    public void setActionBarTitle(String title) {
-        if (title != null)
-            mTitle = title;
-
-        if (getSupportActionBar() != null)
-            getSupportActionBar().setTitle(Html.fromHtml(mTitle));
-    }
-
-    @Override
-    protected void onSaveInstanceState(Bundle outState) {
-        super.onSaveInstanceState(outState);
-        outState.putString(SAVED_TITLE, mTitle);
-    }
-
-    @Override
-    public void onBackStackChanged() {
-        shouldDisplayHomeUp();
-    }
-
-    public void shouldDisplayHomeUp(){
-        //Enable Up button only  if there are entries in the back stack
-        boolean canback = getSupportFragmentManager().getBackStackEntryCount()>0;
-        getSupportActionBar().setDisplayHomeAsUpEnabled(canback);
-    }
-
-    @Override
-    public boolean onSupportNavigateUp() {
-        //This method is called when the up button is pressed. Just the pop back stack.
-        getSupportFragmentManager().popBackStack();
-        return true;
-    }
-}
