--- app-norx/src/main/java/com/alterego/stackoverflow/test/OnFragmentInteractionListener.java
@@ -1,8 +1,0 @@
-package com.alterego.stackoverflow.test;
-
-import android.support.v4.app.Fragment;
-
-public interface OnFragmentInteractionListener {
-    public void setActionBarTitle(String title);
-    public void onRequestOpenFragment(Fragment fragment, String title);
-}
