+++ app-norx/src/main/res/values/servers.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <string name="server">http://api.stackexchange.com/2.2</string>
+    <string name="server">http://api.stackexchange.com/2.2/</string>
 </resources>
