+++ app/build.gradle
@@ -17,7 +17,7 @@ android {
     buildToolsVersion "23.0.3"
 
     defaultConfig {
-        applicationId "com.alterego.stackoverflow.test"
+        applicationId "solutions.alterego.stackoverflow.test"
         minSdkVersion 19
         targetSdkVersion 23
         versionCode 1
@@ -42,6 +42,8 @@ dependencies {
     compile 'com.android.support:design:23.4.0'
     compile 'com.android.support:cardview-v7:23.4.0'
     compile 'com.android.support:recyclerview-v7:23.4.0'
+
+    compile 'solutions.alterego:advancedandroidlogger:1.0.8@aar'
 
     compile 'joda-time:joda-time:2.4'
